requirements.txt
usfm-grammar>=3.1.2
transformers>=4.0.0
wildebeest_nlp>=0.9.2
